Don't raise the severity of the error when there's no digits after a decimal point
Ignore-this: 29a262fc56e15b5719bed96c84072a7f
This causes `onFail` parsers to fail (e.g. it's meant to be either a
Double or a String, but if it starts with a '.' and is not a number,
then `adjustErrBad` will stop it from trying to be parsed as a String).

darcs-hash:20091119023352-97e08-c633a0a8ff163ace082448832518f487060475ce