Be more liberal when parsing statements
Ignore-this: 87123df7a9610815b906fd44442844cb
Up until now, it has been assumed that each statement ended in a `;'.
However, this is not a valid assumption: below is a valid Dot graph:

    digraph { a b -> c -> d e -> f }

This contains 6 nodes and 4 edges.  As such, parsing must be smarter
to be able to tell when statements end.

darcs-hash:20100311121126-97e08-ef80faa79e120a71caf64447f0094ce2a5e3d960