Generalise the types for NodeCluster and ClusterTree
Ignore-this: 1103916c4bea030560f1ca8c8d6ffa75
This is needed for the new GraphvizParams patch to work, but I forgot
to record it; oops!

The rationale for this is so that NodeCluster and ClusterTree can be
used for the intended Data.GraphViz.Algorithms module down the track.

darcs-hash:20100710143151-97e08-15eaa0bc2983c9eb9833343ff72ae547fdfdd5b2