Don't try and be too clever when parsing HTML labels
The attempt at distinguishing between font-based and non-font-based
labels doesn't actually work, as backtracking does not go back to
trying to parse the actual overall FONT attribute.

It's also less efficient in the case of FONT-wrapped Text, as it would
parse it all, realise it doesn't have anything else then - if
backtracking worked - go back and try it all again.

The main reason for this was for error messages anyway.

(The usage of adjustErrBad also prevents backtracking for nested
tags.)

As such, the only part remaining of the commit below that wasn't
reverted was to allow for some extra whitespace.

Revert "Try and improve parsing of HTML labels"

This reverts commit 04f4eab9d19641e8f827070284577cf776897424.